requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
scikit-learn
pandas
numpy
matplotlib

